requirements.txt
openai>=1.13.0
python-dotenv>=1.0.0
agents>=0.1.0
python-telegram-bot>=20.7 